pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chaoschain-governance-os"
version = "0.1.0"
description = "Cross-chain AI-driven governance platform"
readme = "README.md"
authors = [
    {name = "ChaosChain Labs", email = "info@chaoschain.example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "crewai>=0.28.0",
    "web3>=6.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.22.0",
    "alembic>=1.15.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
packages = [
    "agent",
    "agent.crew",
    "agent.crew.researcher",
    "agent.crew.developer", 
    "agent.runtime",
    "agent.tools",
    "verification",
    "verification.tee",
    "verification.attestation",
    "verification.audit",
    "ethereum",
    "ethereum.client",
    "ethereum.monitoring",
    "simulation",
    "simulation.harness",
    "simulation.metrics",
    "simulation.scenarios",
    "reputation",
    "reputation.db",
    "reputation.scoring",
    "reputation.api",
    "api",
    "api.rest",
    "api.models",
    "api.auth"
]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 